non_medical/tracking/DINOv2/requirements.txt
torch
torchvision
opencv-python
scikit-learn
numpy
Pillow
